+++ javascript/components/App.jsx
@@ -1,6 +1,7 @@
 import React, { Component } from 'react'
 import Header from 'components/Header'
 import Menu from 'components/Menu'
+import CanvasContainer from 'components/CanvasContainer'
 
 import './App.css'
 
@@ -15,7 +16,7 @@ export default class App extends Component {
         <Menu />
         <div className='content'>
           <Header />
-          <canvas id='main' />
+          <CanvasContainer />
         </div>
       </div>
     )