+++ javascript/reducers/createReducer.js
@@ -1,7 +1,9 @@
-export default function (name, defaultValue) {
+export default function (name, defaultValue, parentType, parentName) {
   const UPPERCASE_NAME = name.toUpperCase()
 
   return function (state = defaultValue, action) {
+    if (parentName && parentName !== action[parentType]) { return state }
+
     switch (action.type) {
       case `SET_${UPPERCASE_NAME}`:
         return action.value