+++ javascript/utility/configureStore.js
@@ -17,7 +17,8 @@ export default function configureStore () {
 
     saveState({
       fractal: state.fractal,
-      propertiesByFractal: state.propertiesByFractal
+      fractals: state.fractals,
+      viewports: state.viewports
     })
   }, 1000))