+++ shaders/fractal.glsl
@@ -17,7 +17,6 @@ uniform float SUPERSAMPLES;
 
 uniform float COLORSET;
 uniform float FRACTAL;
-uniform float iGlobalTime;
 uniform float EXPONENT;
 
 const int MAX_ITERATIONS = 255;
@@ -109,7 +108,7 @@ vec2 fragCoordToXY(vec4 fragCoord) {
 
   vec2 cartesianPosition = (relativePosition - 0.5) * (X_MAX - X_MIN);
   cartesianPosition.x += center.x;
-  cartesianPosition.y -= center.y;
+  cartesianPosition.y += center.y;
   cartesianPosition.x *= aspectRatio;
 
   return cartesianPosition;